Many many changes. Some of them break the compability, so I made this version 1.6.1 and use Hase server 15 know. However it may be, that I change even more in the future.

+++ server/get_game.php
@@ -19,7 +19,7 @@ while ($row = mysql_fetch_array( $result ))
 {
 	$now = time();
 	$heartbeat_diff = $now - $row['heartbeat_time'];
-	if ($game_row['status'] == 0 && $heartbeat_diff > 90) //one and a half minute no reaction
+	if ($game_row['status'] == 0 && $row['computer'] == 0 && $heartbeat_diff > 90) //one and a half minute no reaction
 	{
 		$player_id = $row['player_id'];
 		$query = "DELETE FROM " . $mysql_prefix . "player_list WHERE game_id = '$game_id' AND player_id = '$player_id'";
